requirements.txt
asgiref==3.7.2
chardet==5.2.0
crispy-bootstrap5==2024.2
cssselect2==0.7.0
Django==4.2.11
django-appconf==1.0.6
django-avatar==8.0.0
django-classy-tags==4.1.0
django-cms==4.1.0
django-crispy-forms==2.1
django-filer==3.1.1
django-formtools==2.5.1
django-js-asset==2.2.0
django-mptt==0.16.0
django-multiselectfield==0.1.12
django-polymorphic==3.1.0
django-ranged-response==0.2.0
django-sekizai==4.1.0
django-simple-captcha==0.6.0
django-treebeard==4.7.1
django-user-agents==0.4.0
djangocms-admin-style==3.3.1
djangocms-attributes-field==3.0.0
djangocms-bootstrap4==3.0.1
djangocms-column==2.0.0
djangocms-file==3.0.1
djangocms-googlemap==2.2.0
djangocms-icon==2.1.0
djangocms-link==3.1.1
djangocms-picture==4.1.1
djangocms-snippet==3.1.1
djangocms-style==3.1.0
djangocms-text-ckeditor==5.1.5
djangocms-video==3.1.0
djangorestframework==3.15.0
dnspython==2.6.1
easy-thumbnails==2.8.5
GDAL==3.7.2 #Change the GDAL version to the one installed on your computer.
gunicorn==21.2.0
html5lib==1.1
lxml==5.1.0
numpy==1.26.4
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
reportlab==4.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
towel==0.31.0
typing_extensions==4.10.0
ua-parser==0.18.0
Unidecode==1.3.8
urllib3==2.2.1
user-agents==2.2.0
webencodings==0.5.1
wincertstore==0.2
XlsxWriter==3.2.0

